+++ Client/src/main/java/cz/cuni/mff/ufal/textan/core/Entity.java
@@ -18,7 +18,7 @@ public class Entity {
     private final int length;
 
     /** Object Type. */
-    private final int type;
+    private final long type;
 
     /** Object candidates for this entity. */
     private final Map<Double, Object> candidates = new HashMap<>();
@@ -34,7 +34,7 @@ public class Entity {
         value = entity.getValue();
         position = entity.getPosition();
         length = entity.getLength();
-        type = (int) entity.getType();
+        type = entity.getType();
     }
 
     /**
@@ -44,7 +44,7 @@ public class Entity {
      * @param length length
      * @param type type
      */
-    public Entity(String value, int position, int length, int type) {
+    public Entity(String value, int position, int length, long type) {
         this.value = value;
         this.position = position;
         this.length = length;
@@ -95,7 +95,7 @@ public class Entity {
      * Returns type.
      * @return type
      */
-    public int getType() {
+    public long getType() {
         return type;
     }
