+++ Client/src/main/java/cz/cuni/mff/ufal/textan/gui/reportwizard/ReportObjectsController.java
@@ -7,25 +7,18 @@ import cz.cuni.mff.ufal.textan.core.processreport.EntityBuilder;
 import cz.cuni.mff.ufal.textan.core.processreport.ProcessReportPipeline;
 import cz.cuni.mff.ufal.textan.core.processreport.Word;
 import cz.cuni.mff.ufal.textan.gui.Utils;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.ResourceBundle;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.geometry.Side;
-import javafx.scene.control.ContextMenu;
-import javafx.scene.control.MenuItem;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.SeparatorMenuItem;
-import javafx.scene.control.Tooltip;
+import javafx.scene.control.*;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.text.Text;
 import javafx.scene.text.TextFlow;
+
+import java.net.URL;
+import java.util.*;
+import java.util.Map.Entry;
 
 /**
  * Controls editing objects.
@@ -88,7 +81,7 @@ public class ReportObjectsController extends ReportWizardController {
         for (final Word word: pipeline.getReportWords()) {
             final Text text = new Text(word.getWord());
             if (word.getEntity() != null) {
-                final int entityId = word.getEntity().getId();
+                final long entityId = word.getEntity().getId();
                 final int entityIndex = word.getEntity().getIndex();
                 text.getStyleClass().add("ENTITY_" + entityId);
